backend/requirements.txt
fastapi
uvicorn[standard]
PyGithub
google-generativeai
python-dotenv
pydantic
httpx
slowapi
groq